modified:   code/src/common/neurons.h 	modified:   code/src/view/NetView.cpp 	modified:   code/src/view/NetView.h 	modified:   code/src/view/utils.cpp 	modified:   code/src/view/utils.h

+++ code/src/view/utils.cpp
@@ -9,26 +9,26 @@ QRectF newQRectF(QPointF center, int radius)
     lower_right.setY(lower_right.y() + radius);
     return QRectF(upper_left, lower_right);
 }
-bool conflict(const QRectF lhs, const QRectF rhs, int radius)
+bool conflict(const QRectF lhs, const QRectF rhs, double radius)
 {
     QPointF center = lhs.center();
     QPointF new_center = rhs.center();
-    int dx = new_center.x() - center.x();
-    int dy = new_center.y() - center.y();
+    double dx = new_center.x() - center.x();
+    double dy = new_center.y() - center.y();
     return (dx * dx + dy * dy <= 4 * radius * radius);
 }
-bool conflict(const QPointF p, const QRectF lhs, int radius)
+bool conflict(const QPointF p, const QRectF lhs, double radius)
 {
     QPointF center = lhs.center();
-    int dx = p.x() - center.x();
-    int dy = p.y() - center.y();
+    double dx = p.x() - center.x();
+    double dy = p.y() - center.y();
     return (dx * dx + dy * dy <= 4 * radius * radius);
 }
-bool isinside(const QPointF p, const QRectF lhs, int radius)
+bool isinside(const QPointF p, const QRectF lhs, double radius)
 {
     QPointF center = lhs.center();
-    int dx = p.x() - center.x();
-    int dy = p.y() - center.y();
+    double dx = p.x() - center.x();
+    double dy = p.y() - center.y();
     return (dx * dx + dy * dy <= radius * radius);
 }
